+++ frontend/src/app/organization/organization-admin/organization-admin-permission.guard.ts
@@ -1,17 +1,13 @@
 import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
 import { map } from 'rxjs';
-import { PermissionService } from './permission.service';
+import { OrganizationAdminPermissionService } from '/workspace/frontend/src/app/organization/organization-admin/organization-admin-permission.service';
 
-// TODO: #4 Allow resource patterns such as role/{id} and replace {id} with _route fragment
-export const permissionGuard = (
-  action: string,
-  resource: string
-): CanActivateFn => {
+export const OrganizationAdminPermissionGuard = (): CanActivateFn => {
   return (_route, _state) => {
-    const permission = inject(PermissionService);
+    const permission = inject(OrganizationAdminPermissionService);
     const router = inject(Router);
-    return permission.check(action, resource).pipe(
+    return permission.checkForOrganizationPermissions().pipe(
       map((isAuthenticated) => {
         if (isAuthenticated) {
           return true;
